% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{iqtn-package}
\alias{iqtn}
\alias{iqtn-package}
\title{iqtn: Interpretable Quantile Temporal Networks}
\description{
Reusable utilities for carbon allowance tail-risk forecasting workflows.
}
\author{
\strong{Maintainer}: First Author \email{author@example.com}

}
