% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{iqtn_spec}
\alias{iqtn_spec}
\title{Create an IQTN model specification}
\usage{
iqtn_spec(
  lookback = 20L,
  horizons = c(1L, 5L, 10L),
  quantile_levels = c(0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9, 0.95, 0.975, 0.99),
  hidden_dim = 64L,
  kernel_size = 3L,
  dilations = c(1L, 2L, 4L),
  dropout = 0.2
)
}
\arguments{
\item{lookback}{Number of trading days in the input sequence.}

\item{horizons}{Forecast horizons.}

\item{quantile_levels}{Quantile levels estimated by the model.}

\item{hidden_dim}{Hidden dimension for the temporal encoder.}

\item{kernel_size}{Temporal convolution kernel size.}

\item{dilations}{Dilation factors for temporal convolution blocks.}

\item{dropout}{Dropout probability.}
}
\value{
A list describing the model specification.
}
\description{
Create an IQTN model specification
}
