% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pinball_loss}
\alias{pinball_loss}
\title{Quantile pinball loss}
\usage{
pinball_loss(y, q, alpha)
}
\arguments{
\item{y}{Numeric vector of realized values.}

\item{q}{Numeric vector of quantile forecasts.}

\item{alpha}{Quantile level in \verb{(0, 1)}.}
}
\value{
Mean pinball loss.
}
\description{
Quantile pinball loss
}
