% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_price_series}
\alias{plot_price_series}
\title{Plot a carbon allowance price series}
\usage{
plot_price_series(
  data,
  date_col = "date",
  price_col = "close",
  main = "Carbon allowance price",
  xlab = "Date",
  ylab = "Price",
  line_col = "#1F6E8C",
  ...
)
}
\arguments{
\item{data}{A data frame containing date and price columns.}

\item{date_col}{Name of the date column.}

\item{price_col}{Name of the price column.}

\item{main}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{line_col}{Line color.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
The plotted data, invisibly.
}
\description{
Plot a carbon allowance price series
}
