% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_warning_timeline}
\alias{plot_warning_timeline}
\title{Plot warning levels over time}
\usage{
plot_warning_timeline(
  data,
  date_col = "date",
  level_col = "warning_level",
  main = "Warning timeline",
  xlab = "Date",
  ylab = "Warning level",
  ...
)
}
\arguments{
\item{data}{A data frame containing date and warning-level columns.}

\item{date_col}{Name of the date column.}

\item{level_col}{Name of the warning-level column. Values should be 0, 1, 2,
and 3.}

\item{main}{Plot title.}

\item{xlab}{X-axis label.}

\item{ylab}{Y-axis label.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
The plotted warning data, invisibly.
}
\description{
Plot warning levels over time
}
