% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{var_exceedance_rate}
\alias{var_exceedance_rate}
\title{VaR exceedance rate}
\usage{
var_exceedance_rate(loss, var)
}
\arguments{
\item{loss}{Numeric vector of realized losses.}

\item{var}{Numeric vector of VaR forecasts.}
}
\value{
Mean exceedance indicator.
}
\description{
VaR exceedance rate
}
